# aria-frontend/openapi/aria.json
{
  "components": {
    "schemas": {
...
      "ApprovalObjectType": {
        "enum": [
          "content_draft",
          "calendar_draft",
          "community_reply",
          "report_draft"
        ],
        "title": "ApprovalObjectType",
        "type": "string"
      },
      "ApprovalQueueResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ContentDraftQueueItem"
                },
                {
                  "$ref": "#/components/schemas/CalendarDraftQueueItem"
                },
                {
                  "$ref": "#/components/schemas/CommunityReplyQueueItem"
                },
                {
                  "$ref": "#/components/schemas/ReportDraftQueueItem"
                }
              ]
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "count",
          "limit",
          "offset"
        ],
        "title": "ApprovalQueueResponse",
        "type": "object"
      },
      "ApprovalResult": {
        "properties": {
          "audit_event": {
            "$ref": "#/components/schemas/ApprovalAuditEvent"
          },
          "decision": {
            "$ref": "#/components/schemas/ApprovalDecision"
          },
          "record": {
            "title": "Record",
            "type": "object"
          }
        },
        "required": [
          "decision",
          "audit_event"
        ],
        "title": "ApprovalResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "main__ApprovalActionRequest": {
        "properties": {
          "metadata": {
            "title": "Metadata",
            "type": "object"
          },
          "object_id": {
            "title": "Object Id",
            "type": "string"
          },
          "object_type": {
            "$ref": "#/components/schemas/ApprovalObjectType"
          },
          "reason": {
            "default": "",
            "title": "Reason",
            "type": "string"
          },
          "requested_changes": {
            "items": {
              "type": "string"
            },
            "title": "Requested Changes",
            "type": "array"
          },
          "reviewer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Id"
          },
          "reviewer_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Role"
          }
        },
        "required": [
          "object_id",
          "object_type"
        ],
        "title": "ApprovalActionRequest",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "ARIA LLM Orchestration Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/internal/ai/approval/approve": {
      "post": {
        "operationId": "ai_approve_draft_internal_ai_approval_approve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/main__ApprovalActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ai Approve Draft"
      }
    },
    "/internal/ai/approval/archive": {
      "post": {
        "operationId": "ai_archive_draft_internal_ai_approval_archive_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/main__ApprovalActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ai Archive Draft"
      }
    },
    "/internal/ai/approval/audit/{object_type}/{object_id}": {
      "get": {
        "operationId": "ai_list_approval_audit_events_internal_ai_approval_audit__object_type___object_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "object_type",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ApprovalObjectType"
            }
          },
          {
            "in": "path",
            "name": "object_id",
            "required": true,
            "schema": {
              "title": "Object Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApprovalAuditEventResponse"
                  },
                  "title": "Response Ai List Approval Audit Events Internal Ai Approval Audit  Object Type   Object Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ai List Approval Audit Events"
      }
    },
    "/internal/ai/approval/decision": {
      "post": {
        "operationId": "ai_apply_approval_decision_internal_ai_approval_decision_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ai Apply Approval Decision"
      }
    },
...
    "/internal/ai/approval/detail/{object_type}/{object_id}": {
      "get": {
        "operationId": "ai_get_approval_detail_internal_ai_approval_detail__object_type___object_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "object_type",
            "required": true,
            "schema": {
              "title": "Object Type",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "object_id",
            "required": true,
            "schema": {
              "title": "Object Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ContentDraftDetail"
                    },
                    {
                      "$ref": "#/components/schemas/CalendarDraftDetail"
                    },
                    {
                      "$ref": "#/components/schemas/CommunityReplyDraftDetail"
                    },
                    {
                      "$ref": "#/components/schemas/ReportDraftDetail"
                    }
                  ],
                  "title": "Response Ai Get Approval Detail Internal Ai Approval Detail  Object Type   Object Id  Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ai Get Approval Detail"
      }
    },
    "/internal/ai/approval/queue": {
      "get": {
        "operationId": "ai_list_approval_queue_internal_ai_approval_queue_get",
        "parameters": [
          {
            "in": "query",
            "name": "brand_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Brand Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "object_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ApprovalObjectType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Object Type"
            }
          },
          {
            "in": "query",
            "name": "platform",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Platform"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "created_after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created After"
            }
          },
          {
            "in": "query",
            "name": "created_before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created Before"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalQueueResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ai List Approval Queue"
      }
    },
...
    "/internal/ai/approval/reject": {
      "post": {
        "operationId": "ai_reject_draft_internal_ai_approval_reject_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/main__ApprovalActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ai Reject Draft"
      }
    },
    "/internal/ai/approval/request-changes": {
      "post": {
        "operationId": "ai_request_changes_internal_ai_approval_request_changes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/main__ApprovalActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ai Request Changes"
      }
    },
    "/internal/ai/approval/submit": {
      "post": {
        "operationId": "ai_submit_for_approval_internal_ai_approval_submit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/main__ApprovalActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ai Submit For Approval"
      }
    },
...
    "/v1/approval/approve": {
      "post": {
        "operationId": "approve_v1_approval_approve_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-ARIA-Workspace-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Workspace-Id"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-Role"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-User-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/api__routers__approval__ApprovalActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Approve",
        "tags": [
          "approval"
        ]
      }
    },
    "/v1/approval/archive": {
      "post": {
        "operationId": "archive_v1_approval_archive_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-ARIA-Workspace-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Workspace-Id"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-Role"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-User-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/api__routers__approval__ApprovalActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Archive",
        "tags": [
          "approval"
        ]
      }
    },
    "/v1/approval/audit/{object_type}/{object_id}": {
      "get": {
        "operationId": "approval_audit_v1_approval_audit__object_type___object_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "object_type",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ApprovalObjectType"
            }
          },
          {
            "in": "path",
            "name": "object_id",
            "required": true,
            "schema": {
              "title": "Object Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Workspace-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Workspace-Id"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-Role"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-User-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApprovalAuditEvent"
                  },
                  "title": "Response Approval Audit V1 Approval Audit  Object Type   Object Id  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Approval Audit",
        "tags": [
          "approval"
        ]
      }
    },
...
  }
}